Force slave host name to address Web UI issues
With the current Mesos web interface the client browser must be able to
communicate directly with the reported slave hostname. In playa-mesos's
case this is a problem, because the VM hostname (mesos.vm) is not
available outside of the VM environment.

This workaround forces the slave to register with a 'hostname'
corresponding to the configurable private IP address.

This currently only works for Mesos 0.15.0+. If we try and set the
hostname argument on 0.14.2 for example, the mesos-slave will fail to
start.