Add option to specify Mesos version
The mesosflexinstall script has support for a '--rel' argument to
specify a Mesos release to be installed. Prior to this commit, if
this argument was passed to the script it would be set, but
wouldn't actually be used when installing Mesos.

This commit introduces a `mesos_release` option in config.json. If
the user creates this configuration option (key), mesosflexinstall
is called with '--rel' and the version string. If the configuration
option doesn't exist in config.json, mesosflexinstall will maintain
its current behavior (install latest).